move DeepState to models
