+++ src/deepresearcher2/models.py
@@ -1,5 +1,18 @@
 #!/usr/bin/env python3
+from __future__ import annotations as _annotations
+
+from dataclasses import dataclass, field
+
 from pydantic import BaseModel, Field
+
+
+@dataclass
+class DeepState:
+    topic: str = "petrichor"
+    search_query: WebSearchQuery | None = field(default_factory=lambda: None)
+    search_results: list[WebSearchResult] | None = field(default_factory=lambda: None)
+    count: int = 0
+    summary: str | None = None
 
 
 class WebSearchQuery(BaseModel):
